Core/Grid: Changed thread based updator to a main thread (world) updater. A constant heart beat is used to call everything and an update queue is used to update maps. The updated queue is set to update 4 maps at once, but it can be set to update from 1 to up to *your choice* maps. Server thread will be used in a near future to update eveything and not only maps.
Solved many multithreading problems and implemented a few new classes.
Next step is to separate everything in its file.

+++ src/Entities/Object.cpp
@@ -41,7 +41,7 @@ Object::~Object()
  *
  * @param diff Time difference from last tick
  */
-bool Object::update(const Poco::UInt32 diff)
+bool Object::update(const Poco::UInt64 diff)
 {
     bool updatedGrid = false;
     if (hasFlag(FLAGS_TYPE_MOVEMENT, FLAG_MOVING))
